added auth in rating and basketball player

+++ src/Controllers/basketballPlayer.ts
@@ -100,10 +100,12 @@ const getBasketballPlayerByID = (req: Request, res: Response, next: NextFunction
 }
 
 const insertNewBasketballPlayer = (req: Request, res: Response, next: NextFunction) => {
-
+    const user = req.user;
     let { basketballPlayer_name, basketballPlayer_position, basketballPlayer_price, basketballPlayer_teamName } = req.body;
-    let query = `INSERT INTO basketball_player (name, position, price, team_name) VALUES ("${basketballPlayer_name}","${basketballPlayer_position}","${basketballPlayer_price}","${basketballPlayer_teamName}")`;
 
+    if(user && user?.roles !== "Admin") return res.status(403).send("No Access");
+
+    let query = `INSERT INTO basketball_player (name, position, price, team_name) VALUES ("${basketballPlayer_name}","${basketballPlayer_position}","${basketballPlayer_price}","${basketballPlayer_teamName}")`;
     ConnectMYSQL()
         .then(connection => {
             Query(connection, query)
@@ -131,10 +133,13 @@ const insertNewBasketballPlayer = (req: Request, res: Response, next: NextFuncti
 }
 
 const deleteBasketballPlayer = (req: Request, res: Response, next: NextFunction) => {
-
+    
+    const user = req.user;
     let playerID = req.params.basketballPlayerID;
-    let query = `DELETE FROM basketball_player WHERE ID = ${playerID}`;
 
+    if(user && user?.roles !== "Admin") return res.status(403).send("No Access");
+
+    let query = `DELETE FROM basketball_player WHERE ID = ${playerID}`;
     ConnectMYSQL()
         .then(connection => {
             Query(connection, query)
@@ -162,11 +167,14 @@ const deleteBasketballPlayer = (req: Request, res: Response, next: NextFunction)
 }
 
 const updateBasketballPlayer = (req: Request, res: Response, next: NextFunction) => {
-
+    
+    const user = req.user;
     let playerID = req.params.basketballPlayerID;
     let { basketballPlayer_name, basketballPlayer_position, basketballPlayer_price, basketballPlayer_teamName } = req.body;
+    
+    if(user && user?.roles !== "Admin") return res.status(403).send("No Access");
+    
     let query = `UPDATE basketball_player  SET name = "${basketballPlayer_name}", position = "${basketballPlayer_position}", price = "${basketballPlayer_price}", team_name = "${basketballPlayer_teamName}" WHERE ID = ${playerID} `;
-
     ConnectMYSQL()
         .then(connection => {
             Query(connection, query)
